requirements.txt
# Core dependencies
anthropic  
fastmcp  
risingwave 
python-dotenv 
asyncio
anyio 
typing-extensions
pandas 
numpy 
psycopg2-binary
datetime
